blonded life

+++ Model/Nota.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 
 namespace Projeto3.Model {
 	public class Nota {
-		public UInt64 Codigo { get; set; }
+		public Int32 Codigo { get; set; }
 		public Cliente Cliente { get; set; }
 		public List<Produto> Produtos { get; set; }
-		public UInt64 Total { get; set; }
+		public Int32 Total { get; set; }
 	}
 }
